Update README.rst
The purpose of this student project at CERN is to rewrite an earlier project, but in a new format that is enduring and permanent. The idea is to be able to open, understand , and use this math project in 50 years with no qualms. The math project is taken from an assignment I was given in 2.g in gymnasium. The assignment is described directly in the code.